# docs/blueprint/openapi.yml
openapi: 3.0.3
info:
  title: Golang Base API
  version: 1.0.0
  description: Starter Kit API documentation
servers:
  - url: http://localhost:3100/api/v1
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          
  /users/me:
    get:
      summary: Get current user profile
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'

  /users:
    get:
      summary: List all users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - role
              properties:
                email:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: User created
          
  /users/{id}:
    put:
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                is_active:
                  type: boolean
      responses:
        '200':
          description: User updated

  /roles/{role}/permissions:
    get:
      summary: Get permissions for a role
      parameters:
        - in: path
          name: role
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: string
    put:
      summary: Update role permissions
      parameters:
        - in: path
          name: role
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Permissions updated

  /permissions/matrix:
    get:
      summary: Get system permission matrix
      responses:
        '200':
          description: System permission matrix
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string

  /users/{id}/permissions:
    get:
      summary: Get user-specific permissions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User permissions map
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties:
                      type: boolean